rename: offer_model
