+++ app/models/offer_model.py
